グループ削除時、もっとも古いグループの画面に遷移するように #38
