+++ backend/package.json
@@ -24,6 +24,7 @@
     "@types/react-modal": "^3.13.1",
     "@types/react-redux": "^7.1.18",
     "@types/react-router-dom": "^5.3.0",
+    "@types/react-router-redux": "^5.0.20",
     "axios": "^0.21",
     "bootstrap": "^4.6.0",
     "browser-sync": "^2.27.5",
@@ -37,6 +38,7 @@
     "postcss": "^8.1.14",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-redux": "^4.0.8",
     "react-test-renderer": "^17.0.2",
     "resolve-url-loader": "4.0.0",
     "sass": "^1.32.11",
